+++ fractalistic/__main__.py
@@ -9,26 +9,61 @@ from .utils import *
 from .app import FractalisticApp
 from . import __version__
 from .colors import color_renderers
+from .vec import Vec
+from .settings import Settings
 from .fractals import fractal_list
 ###### Others
 import gmpy2
 ###############
-  
-
-
 
 @extra_command(params=[])
-@option("-ls", "--load-state", help="Load state from a file", type=STRING)
-@option("-df", "--default-fractal", help="Default fractal to show when starting the program.", type=Choice([frac.__name__ for frac in fractal_list], False), default="Mandelbrot")
-@option("-dc", "--default-color", help="Default color to use when starting the program.", type=Choice([color.__name__ for color in color_renderers], False), default="blue_brown")
-@option("-p", "--decimal-precision", help="Fine-tune numeric precision by specifying the desired bit length for numeric values", type=IntRange(5), default=64)
-@option("-i", "--max-iter", help="This option defines the number of iterations required to classify a point as convergent", type=IntRange(5), default=128)
-@option("-s", "--size", nargs=2, help="Manually set the screenshots width and height. Wont work if you use -f.", type=IntRange(32), default=(1920, 1080))
-@option("-f", "--fit", help="Fit screenshots with the canvas.", is_flag=True)
-@option("-q", "--quality", help="Only when using --fit, set the quality of screenshots by multiplying the original size in the terminal by the specified scaling factor.", type=IntRange(1), default=20)
-@option("-v", "--version", help="Show version number and exit", is_flag=True)
-@option("--debug", help="Enable debug mode for developers.", is_flag=True)
-def main(fit: bool, quality: float, size: tuple[int, int], default_fractal: str, default_color: str, debug: bool, decimal_precision: int, max_iter: int, version: bool, load_state: str):
+@option("-t", "--threads",
+    help="Number of threads to use for rendering", 
+    type=IntRange(1), 
+    default=Settings.threads)
+@option("-st", "--screenshot-threads",
+    help="Number of threads to use for rendering HQ screenshots", 
+    type=IntRange(1), 
+    default=Settings.screenshot_threads)
+@option("-ls", "--load-state", 
+    help="Load state from a file", 
+    type=STRING)
+@option("-df", "--default-fractal", 
+    help="Default fractal to show when starting the program.", 
+    type=Choice([frac.__name__ for frac in fractal_list], False), 
+    default="Mandelbrot")
+@option("-dc", "--default-color", 
+    help="Default color to use when starting the program.", 
+    type=Choice([color.__name__ for color in color_renderers], False), 
+    default="blue_brown")
+@option("-p", "--decimal-precision", 
+    help="Fine-tune numeric precision by specifying the desired bit length for numeric values", 
+    type=IntRange(5),
+    default=Settings.render_settings.wanted_numeric_precision)
+@option("-i", "--max-iter", 
+    help="This option defines the number of iterations required to classify a point as convergent", 
+    type=IntRange(5),
+    default=Settings.render_settings.max_iter)
+@option("-s", "--size", 
+    nargs=2, 
+    help="Manually set the screenshots width and height. Wont work if you use -f.", 
+    type=IntRange(32),
+    default=(Settings.screenshot_size.x, Settings.screenshot_size.y))
+@option("-f", "--fit", 
+    help="Fit screenshots with the canvas.", 
+    is_flag=True)
+@option("-q", "--quality", 
+    help="Only when using --fit, set the quality of screenshots by multiplying the original size in the terminal by the specified scaling factor.", 
+    type=IntRange(1),
+    default=Settings.screenshot_quality)
+@option("-v", "--version", 
+    help="Show version number and exit", 
+    is_flag=True)
+@option("--debug", 
+    help="Enable debug mode for developers.", 
+    is_flag=True)
+
+def main(fit: bool, quality: float, size: tuple[int, int], default_fractal: str, default_color: str, debug: bool, decimal_precision: int, max_iter: int, version: bool, load_state: str, threads: int, screenshot_threads: int):
 
     # If -v or --version is used, show version and exit
     if version:
@@ -38,20 +73,21 @@ def main(fit: bool, quality: float, size: tuple[int, int], default_fractal: str,
     app = FractalisticApp()
 
     # Set default fractal
-    app.fractal_index = get_fractal_index_from_name(default_fractal)
+    app.settings.render_settings.fractal_index = get_fractal_index_from_name(default_fractal)
     
     # Set default color
-    app.color_renderer_index = get_color_index_from_name(default_color)
+    app.settings.render_settings.color_renderer_index = get_color_index_from_name(default_color)
 
-    app.options = {
-        "fit_screenshots": fit,
-        "screenshot_quality": quality,
-        "size": None if size is None else Vec(size[0], size[1]),
-        "max_iter": max_iter,
-        "debug": debug,
-        "numeric_precision": decimal_precision,
-        "state_file": load_state
-    }
+    app.settings.fit_screenshots = fit
+    app.settings.debug = debug
+    app.settings.screenshot_quality = quality
+    app.settings.screenshot_size = Vec(size[0], size[1])
+    app.settings.max_iter = max_iter
+    app.settings.render_settings.wanted_numeric_precision = decimal_precision
+    app.settings.state_file = load_state
+    app.settings.threads = threads
+    app.settings.screenshot_threads = screenshot_threads
+
     app.run()
 
 if __name__ == "__main__":