+++ fractalistic/fractals/julia.py
@@ -1,4 +1,4 @@
-from ..query_config import QueryConfig
+from ..settings import RenderSettings
 from .fractal_base import FractalBase
 from gmpy2 import mpc
 
@@ -10,15 +10,14 @@ class Julia(FractalBase):
     message = "Click a point in the complex plane to generate the corresponding Julia set."
     
     @staticmethod
-    def get(config: QueryConfig) -> int:
+    def get(point: mpc, settings: RenderSettings) -> int:
         i = 0
-        n = config.point_in_plane
 
-        while abs(n) < 4 and i < config.max_iter:
-            n = n*n - config.julia_click
+        while abs(point) < 4 and i < settings.max_iter:
+            point = point*point - settings.julia_click
             i+=1
 
-        if i == config.max_iter:
+        if i == settings.max_iter:
             return -1
         
         return i